INTGR-142: refactor/Correct MyA codeBase file organization
