+++ src/pages/myAdmin/guides/codeBase/usingWithDotnet.tsx
@@ -1,10 +1,9 @@
 import { ReactNode } from "react";
-import { Page } from "../../../components";
-import { PageTitleProps } from "../../../components/PageTitle/PageTitle";
-import { HeaderSections } from "../../../components/Header/headerSectionsEnum";
-import { TableOfContentsItem } from "../../../components/TableOfContents/TableOfContents";
-import Accordion from "../../../components/Accordion/Accordion";
-import { CodeSample } from "../../../components/CodeSamplesContainer";
+import { Page } from "../../../../components";
+import { PageTitleProps } from "../../../../components/PageTitle/PageTitle";
+import { HeaderSections } from "../../../../components/Header/headerSectionsEnum";
+import { TableOfContentsItem } from "../../../../components/TableOfContents/TableOfContents";
+import { CodeSample } from "../../../../components/CodeSamplesContainer";
 
 const overview: ReactNode = (
     <div className="paragraph">
@@ -91,8 +90,8 @@ const pageSections: TableOfContentsItem[] = [
 
 export default function UsingWithDotnet() {
     return (
-        <Page section={HeaderSections.MyAdmin} pageTitle={pageTitle}>
-            {pageSections.map((section) => <Accordion summary={section.summary} p={section.details} id={section.elementId}/>)}
+        <Page section={HeaderSections.MyAdmin} pageTitle={pageTitle} tableOfContents={pageSections}>
+            {null}
         </Page>
     )
 };
